+++ src/components/MaskedInput/MaskedInput.tsx
@@ -1,7 +1,6 @@
 import cx from "classnames";
 import IMask from "imask";
-import { forwardRef, useEffect } from "react";
-import { useIMask } from "react-imask";
+import { forwardRef, useEffect, useRef } from "react";
 import { OmitStrict } from "../../utils/types";
 import { Input, InputProps } from "../Input";
 import { COMPONENT_PREFIX } from "../resources/common.data";
@@ -13,25 +12,35 @@ export type MaskedInputProps = OmitStrict<InputProps, "onChange"> & {
 
 export const MaskedInput = forwardRef<HTMLInputElement, MaskedInputProps>(
   ({ maskOptions, onChange, ...inputProps }, ref) => {
-    const { ref: maskRef, value, unmaskedValue } = useIMask(maskOptions);
+    const inputRef = useRef<HTMLInputElement | null>(null);
+    const iMaskRef = useRef<IMask.InputMask<IMask.AnyMaskedOptions>>();
+
     useEffect(() => {
-      onChange?.(unmaskedValue, value);
-    }, [value, unmaskedValue]);
+      if (inputRef.current) {
+        const iMask = IMask(inputRef.current, maskOptions);
+        iMask.on("accept", () => {
+          onChange?.(iMask.unmaskedValue, iMask.value);
+        });
+
+        iMaskRef.current = iMask;
+
+        return () => {
+          iMask.destroy();
+        };
+      }
+    }, [maskOptions, onChange]);
+
+    useEffect(() => {
+      iMaskRef.current?.updateValue();
+    }, [inputProps.value, inputProps.defaultValue]);
 
     return (
       <Input
         {...inputProps}
         className={cx(`${COMPONENT_PREFIX}-masked-input`, inputProps.className)}
         ref={(newRef) => {
-          if (newRef) {
-            maskRef.current = newRef;
-          }
-
-          if (typeof ref === "function") {
-            ref(newRef);
-          } else if (ref) {
-            ref.current = newRef;
-          }
+          inputRef.current = newRef;
+          ref = inputRef;
         }}
       />
     );