+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2022-05-30
+
+### Added
+
+- ⌚ `DurationPicker`: add duration picker with... no duration yet. Controlled form is under construction!
+- 🔨 Added `getDurationFromMs` and `durationToString` utils
+
 ## [1.1.15] - 2022-05-29
 
 ### Fixed